openai-evaluation-streamlit/pyproject.toml
[tool.poetry]
name = "openai-evaluation-streamlit"
version = "0.1.0"
description = "This package has all the libraries needed to run our code"
authors = ["Dharun Karthick <dharun.ramaraj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
openai = "0.28"
boto3 = "^1.35.19"
pandas = "^2.2.2"
datasets = "^3.0.0"
sqlalchemy = "^2.0.34"
huggingface-hub = "^0.24.7"
python-dotenv = "^1.0.1"
pymssql = "^2.3.1"
matplotlib = "^3.9.2"
python-pptx = "^1.0.2"
mutagen = "^1.47.0"
python-docx = "^1.1.2"
openpyxl = "^3.1.5"
pillow = "^10.4.0"
pypdf2 = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
